Move executor provider from DistributedStreamService to AbstractStreamService and removed useless binding
